Added Dialogue Panel for London and Text
Kámo je skoro půlnoc a já to už nedávám
