+++ PlatBorne/Assets/Scripts/LondonDialogue.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8f37cc77cb204e94c9899109c45e473a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
